Refactor, move validation methods to driver.
