+++ CHANGELOG.md
@@ -8,6 +8,35 @@
   `{didDocument, keyPairs, methodFor}`.
 - **BREAKING**: Remove unused/obsolete `passphrase` parameter.
 
+### Upgrading from <=12.x
+
+**1)** DID Document `generate()` method return signature has changed.
+
+**Before:** `const didDocument = await veresOneDriver.generate();`
+
+The generated `didDocument` was an instance of the `VeresOneDidDoc` class,
+and stored its own generated key pairs in `didDocument.keys`.
+
+**Now:** `const {didDocument, keyPairs, methodFor} = await veresOneDriver.generate();`
+
+In v13, the generated `didDocument` is a plain Javascript object, with no
+methods. Generated keys are returned in the `keyPairs` property (a js `Map`
+instance, with key pairs stored by key id).
+In addition a helper method `methodFor` is provided, to help retrieve keys
+for a particular purpose. For example:
+`methodFor({purpose: 'capabilityInvocation'})` returns the first available
+public/private key pair instance that is referenced in the DID Document's
+`capabilityInvocation` verification relationship.
+
+**X)** Validation methods have changed (used by the `did-veres-one` validator
+node):
+
+- `didDocument.validateDid({mode})` becomes:
+  `VeresOneDriver.validateDid({didDocument, mode})`
+- `didDocument.validateMethodIds()` becomes:
+  `VeresOneDriver.validateMethodIds({didDocument})`
+
+
 ## 13.0.0 - 2021-03-12
 
 ### Changed
