+++ lib/VeresOneClient.js
@@ -3,13 +3,13 @@
  */
 'use strict';
 
-const {create} = require('apisauce');
 const base64url = require('base64url-universal');
+const {httpClient} = require('@digitalbazaar/http-client');
 const httpSignatureHeader = require('http-signature-header');
 const jsonld = require('jsonld');
 const {WebLedgerClient} = require('web-ledger-client');
 const VeresOneClientError = require('./VeresOneClientError');
-const VeresOneDidDoc = require('./VeresOneDidDoc');
+const {fromNym} = require('./VeresOneDriver');
 jsonld.documentLoader = require('./documentLoader');
 
 const {createAuthzHeader, createSignatureString} = httpSignatureHeader;
@@ -19,20 +19,21 @@ const {createAuthzHeader, createSignatureString} = httpSignatureHeader;
  */
 class VeresOneClient {
   /**
-   * @param hostname {string} Hostname of the ledger (points to a load balancer
-   *   or a specific node).
-   * @param ledger {WebLedgerClient}
+   * @param {object} options - Options hashmap.
+   * @param {string} options.hostname - Hostname of the ledger (points to a load
+   *   balancer or a specific node).
+   * @param {WebLedgerClient} [ledger] - Web Ledger Client instance.
 
-   * @param mode {string} One of 'dev'/'test'/'live'. Determines https
-   *   agent settings (does not reject unsigned certs in 'dev' mode, for
-   *   example).
-   * @param [httpsAgent] {Agent} A NodeJS HTTPS Agent (`https.Agent`) instance.
-   * @param [logger]
+   * @param {string} [mode] - One of 'dev'/'test'/'live'.
+   * @param {Agent} [httpsAgent] - A NodeJS HTTPS Agent (`https.Agent`)
+   *   instance.
+   * @param {object} [logger=console] - Logger instance (with .log(), warn() and
+   *   error() methods).
    */
   constructor({hostname, ledger, mode, httpsAgent, logger = console}) {
     this.hostname = hostname;
     if(!hostname) {
-      throw new TypeError('Missing ledger hostname.');
+      throw new TypeError('The "hostname" parameter is required.');
     }
 
     this.ledger = ledger || new WebLedgerClient({httpsAgent, hostname, logger});
@@ -45,8 +46,8 @@ class VeresOneClient {
    * Fetches a DID Document for a given DID. If it contains a #hash fragment,
    * it's likely a key id, so just return the subgraph, not the full doc.
    *
-   * @param did {string} DID uri (possibly with hash fragment)
-   * @param forceConstruct {boolean} Forces deterministic construction of
+   * @param {string} did - DID uri (possibly with hash fragment)
+   * @param {boolean} forceConstruct - Forces deterministic construction of
    *   DID Document from cryptonym.
    *
    * @returns {Promise<object>} Resolves to DID Document Fetch Result
@@ -63,14 +64,14 @@ class VeresOneClient {
 
     // FIXME - remove, replace with cache
     if(isNym && forceConstruct) {
-      didDoc = (await VeresOneDidDoc.fromNym({did: docUri})).doc;
+      didDoc = (await fromNym({did: docUri})).doc;
     } else {
       try {
         result = await this.ledger.getRecord({id: docUri});
         didDoc = result.record;
       } catch(error) {
         if(error.name === 'NotFoundError' && isNym) {
-          didDoc = (await VeresOneDidDoc.fromNym({did: docUri})).doc;
+          didDoc = (await fromNym({did: docUri})).doc;
         } else {
           throw error;
         }
