+++ lib/VeresOneDidDoc.js
@@ -8,7 +8,6 @@ const constants = require('./constants');
 const jsonpatch = require('fast-json-patch');
 const jsonld = require('jsonld');
 
-const VERES_DID_REGEX = /^(did:v1:)(test:)?(uuid|nym):(.+)/;
 const didContexts = [
   constants.DID_CONTEXT_URL,
   constants.VERES_ONE_CONTEXT_URL
@@ -23,187 +22,6 @@ class VeresOneDidDoc {
 
     // Includes private keys -- this property will not be serialized.
     this.keys = options.keys || {};
-  }
-
-  /**
-   * Creates a DID Document from a cryptonym DID.
-   *
-   * @param {string} did
-   * @returns {Promise<VeresOneDidDoc>}
-   */
-  static async fromNym({did}) {
-    let invokeKey, mode;
-    try {
-      // Re-hydrate capabilityInvocation public key from fingerprint
-      const match = VERES_DID_REGEX.exec(did);
-      mode = match[2] && match[2].slice(0, -1); // 'test' or otherwise
-      const fingerprint = match[4];
-      invokeKey = LDKeyPair.fromFingerprint({fingerprint});
-    } catch(error) {
-      throw new Error(`Invalid cryptonym: ${did}`);
-    }
-    // Use that key to deterministically construct a DID Doc
-    return VeresOneDidDoc.generate({
-      mode,
-      keyType: invokeKey.type,
-      invokeKey,
-      authKey: invokeKey,
-      delegateKey: invokeKey,
-      assertionKey: invokeKey
-    });
-  }
-
-  /**
-   * Validates the DID of this document.
-   * - Ensures DID contains 'test:' when running in 'test' mode and vice versa
-   * - If cryptonym DIDs, ensures nym is validated against the invocation key
-   * - Tests for invalid characters in the Specific ID
-   *
-   * @see https://w3c-ccg.github.io/did-spec/#the-generic-did-scheme
-   *
-   * @param [mode='dev'] {string} Mode: 'test'/'live' etc
-   *
-   * @returns {Promise} validator report
-   *   {boolean} valid - true if the DID is valid
-   *   {Error} error - included when `valid` is false
-   */
-  async validateDid({mode = constants.DEFAULT_MODE} = {}) {
-    // Make an exception for urn:uuid: type DID (elector pool doc, for example)
-    if(this.id && this.id.startsWith('urn:uuid:')) {
-      return {valid: true}; // short-circuit, UUID URNs are fine as is
-    }
-
-    const match = VERES_DID_REGEX.exec(this.id);
-
-    if(!match) {
-      return {
-        error: new Error(`Invalid DID format: "${this.id}".`),
-        valid: false
-      };
-    }
-
-    // [2] undefined or 'test:'
-    const didMode = match[2] && match[2].slice(0, -1);
-    const type = match[3]; // nym / uuid
-    const id = match[4];
-
-    if(mode === 'test' && didMode !== 'test') {
-      return {
-        error: new Error(`DID is invalid for test mode: "${this.id}".`),
-        valid: false
-      };
-    }
-
-    if(mode !== 'test' && didMode === 'test') {
-      return {
-        error: new Error(
-          `Test DID does not match mode "${mode}": "${this.id}".`),
-        valid: false
-      };
-    }
-
-    // ensure no invalid characters
-    if((/[^A-Za-z0-9:\-.]+/).exec(id)) {
-      return {
-        error: new Error(
-          `Specific id contains invalid characters: "${this.id}".`),
-        valid: false
-      };
-    }
-
-    // if type is 'uuid', no further validation necessary at the moment
-
-    if(type === 'nym') {
-      return this.validateCryptonymDid();
-    }
-
-    // success
-    return {valid: true};
-  }
-
-  /**
-   * Validates the (nym-based) DID of this document against invocation key.
-   *
-   * Note: Only validates the 'nym' part of the DID, assumes the overall
-   * format was validated already (by `validateDid()`).
-   *
-   * @returns {Promise} validator report
-   *   {boolean} valid - true if the DID is valid
-   *   {Error} error - included when `valid` is false
-   */
-  async validateCryptonymDid() {
-    if(!this.doc.capabilityInvocation) {
-      return {
-        error: new Error('Cryptonym DID requires a capabilityInvocation key.'),
-        valid: false
-      };
-    }
-
-    const keyPair = await LDKeyPair.from(
-      this.getVerificationMethod({proofPurpose: 'capabilityInvocation'}));
-
-    if(!keyPair || !keyPair.publicKey) {
-      return {
-        error: new Error('Public key is required for cryptonym verification.'),
-        valid: false
-      };
-    }
-
-    const fingerprint = VERES_DID_REGEX.exec(this.id)[4];
-
-    // verifyFingerprint has the same validator return signature
-    return keyPair.verifyFingerprint(fingerprint);
-  }
-
-  /**
-   * Validates the method IDs of this document. Method IDs must be of the
-   * format: did#<multibase fingerprint>
-   *
-   * @returns {Promise} validator report
-   *   {boolean} valid - true if the DID is valid
-   *   {Error} error - included when `valid` is false
-   */
-  async validateMethodIds() {
-    for(const proofPurpose in constants.PROOF_PURPOSES) {
-      const methods = this.getAllVerificationMethods(proofPurpose);
-      if(!methods) {
-        // This DID document does not contain any methods for this purpose
-        continue;
-      }
-      for(const method of methods) {
-        // TODO: support methods that are not LDKeyPairs
-        const keyPair = await LDKeyPair.from(method);
-        // note: Veres One DID documents presently do not permit keys from
-        // other DID documents (or other HTTPS resources, etc)
-        const parts = keyPair.id.split('#');
-        if(parts.length !== 2) {
-          return {
-            error: new Error(
-              'Invalid DID key ID; key ID must be of the form ' +
-              '"<did>#<multibase key fingerprint>".'),
-            valid: false
-          };
-        }
-
-        if(parts[0] !== this.id) {
-          return {
-            error: new Error(
-              'Invalid DID key ID; key ID does not match the DID.'),
-            valid: false
-          };
-        }
-
-        const fingerprint = parts[1];
-
-        // verifyFingerprint has the same validator return signature
-        const result = keyPair.verifyFingerprint(fingerprint);
-        if(!result.valid) {
-          return result;
-        }
-      }
-    }
-    // success
-    return {valid: true};
   }
 
   /**
