+++ lib/VeresOneDriver.js
@@ -4,19 +4,14 @@
 'use strict';
 
 const didIo = require('@digitalbazaar/did-io');
-const {CapabilityInvocation} = require('@digitalbazaar/zcapld');
 const {Ed25519Signature2020} = require('@digitalbazaar/ed25519-signature-2020');
-const jsigs = require('jsonld-signatures');
 const {Ed25519VerificationKey2020} =
   require('@digitalbazaar/ed25519-verification-key-2020');
 const {X25519KeyAgreementKey2020} =
   require('@digitalbazaar/x25519-key-agreement-key-2020');
 const uuid = require('uuid-random');
-
 const constants = require('./constants');
-const documentLoader = require('./documentLoader');
 const veresOneContext = require('veres-one-context');
-const VeresOneDidDoc = require('./VeresOneDidDoc');
 const VeresOneClient = require('./VeresOneClient');
 
 const {CryptoLD} = require('crypto-ld');
@@ -26,6 +21,8 @@ cryptoLd.use(X25519KeyAgreementKey2020);
 
 const {DEFAULT_DID_TYPE, DEFAULT_KEY_TYPE, DEFAULT_MODE} = constants;
 
+const DID_REGEX = /^(did:v1:)(test:)?(uuid|nym):(.+)/;
+
 const DID_DOC_CONTEXTS = [
   'https://www.w3.org/ns/did/v1',
   'https://w3id.org/veres-one/v1',
@@ -33,11 +30,9 @@ const DID_DOC_CONTEXTS = [
   'https://w3id.org/security/suites/x25519-2020/v1'
 ];
 
-const DID_REGEX = /^(did:v1:)(test:)?(uuid|nym):(.+)/;
-
 class VeresOneDriver {
   /**
-   * @param [options={}] {object} - Options hashmap
+   * @param [options={}] {object} - Options hashmap.
    * @param {string} [options.mode='test'] Ledger mode ('test', 'dev', 'live'),
    *   determines hostname for ledger client.
    * @param {string} [options.hostname] Optional hostname override. If not
@@ -53,8 +48,7 @@ class VeresOneDriver {
 
     this.logger = logger || console;
 
-    this.hostname = hostname ||
-      VeresOneDriver.defaultHostname({mode: this.mode});
+    this.hostname = hostname || _defaultHostname({mode: this.mode});
 
     this.client = client ||
       new VeresOneClient({
@@ -63,78 +57,6 @@ class VeresOneDriver {
         mode: this.mode,
         logger: this.logger
       });
-  }
-
-  /**
-   * @returns {string} Hostname for current mode (dev/live etc)
-   */
-  static defaultHostname({mode} = {}) {
-    switch(mode) {
-      case 'dev':
-        return 'node-1.veres.one.local:45443';
-      case 'test':
-        return 'ashburn.capybara.veres.one';
-      case 'live':
-        return 'veres.one';
-      default:
-        throw new Error(`Unknown mode: "${mode}".`);
-    }
-  }
-
-  /**
-   * Attaches proofs to an operation by:
-   *
-   *  1. Using an Accelerator service, in which case an authorization DID
-   *     Document is required beforehand (typically obtained in exchange for
-   *     payment).
-   *
-   * @param operation {object} WebLedger operation
-   *
-   * @param options {object}
-   *
-   * @param [options.accelerator] {string} Hostname of accelerator to use
-   * @param [options.authDoc] {VeresOneDidDoc} Auth DID Doc, required if using
-   *   an accelerator service
-   *
-   * @param [options.notes]
-   *
-   * @returns {Promise<Operation>} an operation document with proofs attached.
-   */
-  async attachProofs({operation, options}) {
-    const {didDocument} = options;
-
-    if(options.accelerator) {
-      // send operation to an accelerator for proof
-      this.logger.log('Sending to accelerator for proof:', options.accelerator);
-      operation = await this.attachAcceleratorProof({operation, ...options});
-    } else {
-      // send to ticket service for a proof
-      operation = await this.attachTicketServiceProof({operation});
-    }
-
-    // get the capability invocation key, for signing the proof
-    const invokeKeyNode = didDocument.getVerificationMethod({
-      proofPurpose: 'capabilityInvocation'
-    });
-    const creator = invokeKeyNode.id;
-    const invokeKey = didDocument.keys[invokeKeyNode.id];
-    if(!invokeKey || !invokeKey.privateKey) {
-      throw new Error('Invocation key required to perform a send.');
-    }
-
-    // attach capability invocation proof
-    const capabilityAction =
-      operation.type.startsWith('Create') ? 'create' : 'update';
-
-    operation = await this.attachInvocationProof({
-      operation,
-      capability: didDocument.id,
-      capabilityAction,
-      creator,
-      key: invokeKey
-    });
-
-    return operation;
   }
 
   /**
@@ -191,19 +113,10 @@ class VeresOneDriver {
   }
 
   /**
-   * Generates a new DID Document.
+   * Generates a new DID Document. (See static `generate()` docstring for
+   * details).
    *
-   * @param [didType='nym'] {string} DID type, 'nym' or 'uuid'
-   * @param [keyType] {string}
-   * @param [passphrase] {string}
-   * @param [invokeKey] {LDKeyPair} Optional invocation key to serve as the DID
-   *   basis (useful if you've generated a key via a KMS).
-   * @param [authKey] {LDKeyPair} Optional
-   * @param [delegateKey] {LDKeyPair} Optional
-   * @param [assertionKey] {LDKeyPair} Optional
-   * @param [keyAgreementKey] {LDKeyPair} Optional
-   *
-   * @returns {Promise<{didDocument: object, keyPairs: Map,
+   * @returns {Promise<{didDocument: Object, keyPairs: Map,
    *   methodFor: Function}>} Resolves with the generated DID Document, along
    *   with the corresponding key pairs used to generate it.
    */
@@ -213,6 +126,38 @@ class VeresOneDriver {
     keyAgreementKey
   } = {}) {
     const {mode} = this;
+    return VeresOneDriver.generate({
+      didType, keyType, mode, invokeKey, authKey, delegateKey, assertionKey,
+      keyAgreementKey
+    });
+  }
+
+  /**
+   * Generates a new DID Document.
+   *
+   * @param {object} [options={}] - Options hashmap.
+   * @param {string} [options.didType='nym'] - DID type, 'nym' or 'uuid'.
+   * @param {string} [options.keyType=DEFAULT_KEY_TYPE] - Verification key type.
+   *
+   * The following keys are optional, and will be generated if not passed in.
+   *
+   * @param {LDKeyPair} [options.invokeKey] - Capability invocation key pair
+   *   (useful if you've generated a key via a KMS). If present, used to
+   *   deterministically derive a cryptonym DID.
+   * @param {LDKeyPair} [options.authKey] - Authentication key pair.
+   * @param {LDKeyPair} [options.delegateKey] - Capability delegation key pair.
+   * @param {LDKeyPair} [options.assertionKey] - Assertion method key pair.
+   * @param {LDKeyPair} [options.keyAgreementKey] - Key agreement key pair.
+   *
+   * @returns {Promise<{didDocument: object, keyPairs: Map,
+   *   methodFor: Function}>} Resolves with the generated DID Document, along
+   *   with the corresponding key pairs used to generate it.
+   */
+  static async generate({
+    didType = DEFAULT_DID_TYPE, keyType = DEFAULT_KEY_TYPE, mode,
+    invokeKey, authKey, delegateKey, assertionKey,
+    keyAgreementKey
+  } = {}) {
     const keyPairs = new Map();
 
     // Before we initialize the rest of the keys, we need to compose the DID
@@ -374,76 +319,167 @@ class VeresOneDriver {
   }
 
   /**
-   * Sends a ledger operation to an accelerator.
-   * Required when registering a DID Document (and not using a proof of work).
+   * Validates the DID of this document.
+   * Used by the `veres-one-validator` node.
    *
-   * @param options {object}
+   * - Ensures DID contains 'test:' when running in 'test' mode and vice versa
+   * - If cryptonym DIDs, ensures nym is validated against the invocation key
+   * - Tests for invalid characters in the Specific ID
    *
-   * @returns {Promise<object>} Response from an axios POST request
+   * @param {object} options - Options hashmap.
+   * @param {string} didDocument - DID document to validate.
+   * @param {string} [mode='dev'] - Mode: 'test'/'live' etc
+   *
+   * @returns {Promise<{valid: boolean, error: Error}>} - Resolves with the
+   *   validation result.
    */
-  async attachAcceleratorProof(options) {
-    let authKey;
-
-    try {
-      authKey = options.authDoc.getVerificationMethod(
-        {proofPurpose: 'authentication'});
-    } catch(error) {
-      throw new Error('Missing or invalid Authorization DID Doc.');
+  static async validateDid({didDocument, mode = constants.DEFAULT_MODE} = {}) {
+    if(!(didDocument && didDocument.id)) {
+      throw new TypeError('The "didDocument.id" parameter is required.');
+    }
+    const did = didDocument.id;
+    if(typeof did !== 'string') {
+      return {
+        valid: false,
+        error: new Error('DID must be a string.')
+      };
+    }
+    // Make an exception for urn:uuid: type DID (elector pool doc, for example)
+    if(did.startsWith('urn:uuid:')) {
+      return {valid: true}; // short-circuit, UUID URNs are fine as is
     }
 
-    // send DID Document to a Veres One accelerator
-    this.logger.log('Generating accelerator signature...');
-    return this.client.sendToAccelerator({
-      operation: options.operation,
-      hostname: options.accelerator,
-      env: options.mode,
-      authKey
-    });
+    const match = DID_REGEX.exec(did);
+
+    if(!match) {
+      return {
+        error: new Error(`Invalid DID format: "${did}".`),
+        valid: false
+      };
+    }
+
+    // [2] undefined or 'test:'
+    const didMode = match[2] && match[2].slice(0, -1);
+    const type = match[3]; // nym / uuid
+    const id = match[4];
+
+    if(mode === 'test' && didMode !== 'test') {
+      return {
+        error: new Error(`DID is invalid for test mode: "${did}".`),
+        valid: false
+      };
+    }
+
+    if(mode !== 'test' && didMode === 'test') {
+      return {
+        error: new Error(
+          `Test DID does not match mode "${mode}": "${did}".`),
+        valid: false
+      };
+    }
+
+    // ensure no invalid characters
+    if((/[^A-Za-z0-9:\-.]+/).exec(id)) {
+      return {
+        error: new Error(
+          `Specific id contains invalid characters: "${did}".`),
+        valid: false
+      };
+    }
+
+    // if type is 'uuid', no further validation necessary at the moment
+
+    if(type === 'nym') {
+      return this._validateCryptonymDid({didDocument});
+    }
+
+    // success
+    return {valid: true};
   }
 
   /**
-   * Adds an ocap invocation proof to an operation.
+   * Validates the (nym-based) DID of this document against invocation key.
    *
-   * @param {string} capability - capability url (did)
-   * @param {string} capabilityAction - Here, 'create' or 'update'
-   * @param {object} operation - WebLedger operation result (either from
-   *   `attachAcceleratorProof()` or `attachTicketServiceProof()`)
-   * @param {Ed25519KeyPair} key - invocation key
+   * Note: Only validates the 'nym' part of the DID, assumes the overall
+   * format was validated already (by `validateDid()`).
    *
-   * @returns {Promise<object>}
+   * @returns {Promise<{valid: boolean, error: Error}>} - Resolves with the
+   *   validation result.
    */
-  async attachInvocationProof({capability, capabilityAction, operation, key}) {
-    return jsigs.sign(operation, {
-      documentLoader,
-      compactProof: false,
-      suite: new Ed25519Signature2020({key}),
-      purpose: new CapabilityInvocation({capability, capabilityAction})
+  static async _validateCryptonymDid({didDocument} = {}) {
+    const did = didDocument.id;
+    const capabilityInvocationMethod = didIo.findVerificationMethod({
+      doc: didDocument, purpose: 'capabilityInvocation'
     });
+    if(!capabilityInvocationMethod) {
+      return {
+        error: new Error('Cryptonym DID requires a capabilityInvocation key.'),
+        valid: false
+      };
+    }
+    const keyPair = await Ed25519Signature2020.from(capabilityInvocationMethod);
+    if(!(keyPair && keyPair.publicKeyMultibase)) {
+      return {
+        error: new Error('Public key is required for cryptonym verification.'),
+        valid: false
+      };
+    }
+    const fingerprint = DID_REGEX.exec(did)[4];
+
+    // verifyFingerprint has the desired validator return signature
+    return keyPair.verifyFingerprint({fingerprint});
   }
 
   /**
-   * Adds a zcap delegation proof to a capability DID Document.
+   * Validates the method IDs of this document. Method IDs must be of the
+   * format: did#<multibase fingerprint>
+   * Used by `veres-one-validator` nodes.
+   *
+   * @returns {Promise} validator report
+   *   {boolean} valid - true if the DID is valid
+   *   {Error} error - included when `valid` is false
    */
-  async attachDelegationProof({didDocument, creator, privateKeyPem}) {
-    // FIXME: validate didDocument, creator, and privateKeyPem
-    // TODO: support `signer` API as alternative to `privateKeyPem`
-    return jsigs.sign(didDocument.doc, {
-      algorithm: 'RsaSignature2018',
-      creator,
-      privateKeyPem,
-      proof: {
-        '@context': constants.VERES_ONE_CONTEXT_URL,
-        proofPurpose: 'capabilityDelegation'
+  async validateMethodIds() {
+    for(const proofPurpose in constants.VERIFICATION_RELATIONSHIPS) {
+      const methods = this.getAllVerificationMethods(proofPurpose);
+      if(!methods) {
+        // This DID document does not contain any methods for this purpose
+        continue;
       }
-    });
-  }
+      for(const method of methods) {
+        // TODO: support methods that are not LDKeyPairs
+        const keyPair = await LDKeyPair.from(method);
+        // note: Veres One DID documents presently do not permit keys from
+        // other DID documents (or other HTTPS resources, etc)
+        const parts = keyPair.id.split('#');
+        if(parts.length !== 2) {
+          return {
+            error: new Error(
+              'Invalid DID key ID; key ID must be of the form ' +
+              '"<did>#<multibase key fingerprint>".'),
+            valid: false
+          };
+        }
 
-  async attachTicketServiceProof({operation}) {
-    const s = await this.client.getStatus();
-    const ticketService = s.service['urn:veresone:ticket-service'].id;
-    const result = await this.client.getTicketServiceProof(
-      {operation, ticketService});
-    return result.operation;
+        if(parts[0] !== did) {
+          return {
+            error: new Error(
+              'Invalid DID key ID; key ID does not match the DID.'),
+            valid: false
+          };
+        }
+
+        const fingerprint = parts[1];
+
+        // verifyFingerprint has the same validator return signature
+        const result = keyPair.verifyFingerprint(fingerprint);
+        if(!result.valid) {
+          return result;
+        }
+      }
+    }
+    // success
+    return {valid: true};
   }
 }
 
@@ -451,6 +487,42 @@ VeresOneDriver.contexts = {
   [constants.VERES_ONE_CONTEXT_URL]:
     veresOneContext.contexts.get(constants.VERES_ONE_CONTEXT_URL)
 };
+
+/**
+ * Creates a DID Document from a cryptonym DID.
+ * (This is very similar to how a `did:key` DID document is created.)
+ *
+ * @param {string} did - Cryptonym DID to re-hydrate into a did document.
+ *
+ * @returns {Promise<{didDocument: object, keyPairs: Map,
+ *   methodFor: Function}>} Resolves with the generated DID Document.
+ */
+async function fromNym({did} = {}) {
+  if(!did) {
+    throw new TypeError('The "did" parameter is required.');
+  }
+  let invokeKey;
+  let mode;
+  try {
+    // Re-hydrate capabilityInvocation public key from fingerprint
+    const match = DID_REGEX.exec(did);
+    mode = match[2] && match[2].slice(0, -1); // 'test' or otherwise
+    const fingerprint = match[4];
+    invokeKey = Ed25519Signature2020.fromFingerprint({fingerprint});
+  } catch(error) {
+    throw new Error(`Invalid cryptonym: ${did}`);
+  }
+  // Use that key to deterministically construct a DID Doc
+  return VeresOneDriver.generate({
+    didType: 'nym',
+    mode,
+    keyType: invokeKey.type,
+    invokeKey,
+    authKey: invokeKey,
+    delegateKey: invokeKey,
+    assertionKey: invokeKey
+  });
+}
 
 /**
  * Generates a DID uri, either as a globally unique random string (uuid),
@@ -493,7 +565,24 @@ function _createCryptonymDid({key, mode = constants.DEFAULT_MODE}) {
   return `${prefix}:nym:${key.fingerprint()}`;
 }
 
+/**
+ * @returns {string} Hostname for current mode (dev/live etc)
+ */
+function _defaultHostname({mode} = {}) {
+  switch(mode) {
+    case 'dev':
+      return 'node-1.veres.one.local:45443';
+    case 'test':
+      return 'ashburn.capybara.veres.one';
+    case 'live':
+      return 'veres.one';
+    default:
+      throw new Error(`Unknown mode: "${mode}".`);
+  }
+}
+
 module.exports = {
   VeresOneDriver,
+  fromNym,
   DID_REGEX
 };
